+++ src/ckzg.c
@@ -18,10 +18,8 @@
 #include "common/bytes.c"
 #include "common/ec.c"
 #include "common/fr.c"
-#include "common/helpers.c"
 #include "common/lincomb.c"
-#include "common/ret.c"
-#include "common/settings.c"
+#include "common/utils.c"
 #include "eip4844/api.c"
 #include "eip4844/blob.c"
 #include "eip7594/api.c"
