+++ src/eip7594/api.c
@@ -18,8 +18,8 @@
 
 #include "common/alloc.h"
 #include "common/fr.h"
-#include "common/helpers.h"
 #include "common/lincomb.h"
+#include "common/utils.h"
 #include "eip7594/fft.h"
 
 #include <assert.h> /* For assert */
