+++ src/eip7594/fft.c
@@ -15,7 +15,7 @@
  */
 
 #include "eip7594/fft.h"
-#include "common/helpers.h"
+#include "common/utils.h"
 #include "eip4844/blob.h"
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
