+++ src/setup/api.c
@@ -16,7 +16,7 @@
 
 #include "setup/api.h"
 #include "common/alloc.h"
-#include "common/helpers.h"
+#include "common/utils.h"
 #include "eip7594/api.h"
 #include "eip7594/fft.h"
